remove duplicate dummy client impl classes
